Conflicts:
	Cat-apps/src/com/hogg/catapps/cat/Cat.java
	Cat-apps/src/com/hogg/catapps/ui/meter/HungerMeter.java
	Cat-apps/src/com/hogg/catapps/ui/meter/ThirstMeter.java

Changed default hunger and thirst value to be 50.

+++ Cat-apps/src/com/hogg/catapps/ui/meter/HungerMeter.java
@@ -9,6 +9,7 @@
  * Contributors:
  * 		James Garner
  * 		Drew Burden
+ * 		Daniel Thomas
  * 
  * Copyright © 2012 Hogg Studios
  * All rights reserved.
@@ -22,7 +23,7 @@ public class HungerMeter extends Meter {
 	// by moodPercentageVariable% of the incrementAmount
 	double moodPercentageVariable = 0.10; 
 	
-	public HungerMeter(int initValue, int incAmount) {
+	public HungerMeter(int initValue, double incAmount) {
 		super(initValue, incAmount);
 	}
 }
